requirements.txt
pyyaml
requests
paho-mqtt
gevent
qhue
influxdb